Pull out a shared requestHelper for setting headers
