+++ lib/v1_api_calls/memberships.js
@@ -1,16 +1,11 @@
 var stampit = require('stampit');
-var request = require('superagent-bluebird-promise');
-
-var CLUSTER_ID_HEADER = 'X-Giant-Swarm-ClusterID';
+var RequestHelper = require('../request_helper.js');
 
 module.exports = stampit().
+  compose(RequestHelper).
   methods({
     memberships: function() {
-      headers = {};
-      headers[CLUSTER_ID_HEADER] = this.clusterId;
-
-      return request.get(this.apiEndpoint + "/v1/user/me/memberships").
-      set(headers).
+      return this.getRequest(this.apiEndpoint + "/v1/user/me/memberships").
       then(function(response) {
         return {
           result: response.body.data,
