+++ lib/v1_api_calls/ping.js
@@ -1,5 +1,5 @@
 var stampit = require('stampit');
-var request = require('superagent-bluebird-promise');
+var RequestHelper = require('../request_helper.js');
 
 // Returns true when the server is up
 // Returns false when the server is down
@@ -7,7 +7,7 @@ var request = require('superagent-bluebird-promise');
 module.exports = stampit().
   methods({
     ping: function() {
-      return request.get(this.apiEndpoint + "/v1/ping").
+      return this.getRequest(this.apiEndpoint + "/v1/ping").
       then(function(response) {
         if (response.text === '"OK"') {
           return {
